pyproject.toml
[tool.poetry]
name = "dukkha"
version = "0.1.0"
description = ""
authors = ["Casey Johnson <caseyjohnsonwv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.43.0"
langchain = "^0.2.16"
langchain-community = "^0.2.16"
faiss-cpu = "^1.8.0.post1"
python-dotenv = "^1.0.1"
pypdf = "^4.3.1"
gradio = "^4.42.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
